// src/app/shared/models/flight.model.ts
export class Flight {
    idFlight: number;
    companyName: string;
    flightType: string;
    travelType: string;
    departureDate: Date;
    departureTime: Date;
    arrivalDate: Date;
    arrivalTime: Date;
    backDate: Date;
    backTime: Date;
    departureLocation: string;
    arrivalLocation: string;
    flightDuration: Date;
    connectionDuration: Date;
    aircraftType: string;
}